+++ services/bot_services.py
@@ -11,6 +11,7 @@ from typing import TYPE_CHECKING
 
 from services.app_settings import AppSettings
 from services.application.frog_limit_service import FrogRateLimiterService
+from services.application.frog_requests import FrogRequestService
 from services.application.image_service import ImageService
 
 if TYPE_CHECKING:
@@ -43,5 +44,6 @@ class BotServices:
     settings: AppSettings
     image_service: ImageService
     frog_rate_limiter: FrogRateLimiterService
+    frog_request_service: FrogRequestService
     scheduler: TaskScheduler | None = None
     bot_controller: WednesdayBot | None = None  # для команд управления ботом, например /stop